Split cell32 in two
mksh uses signed 32 bits integers for its calculation and segfaults on the
integer underflow.

bash uses signed 64 bits integers and thus handle the integer underflow
very well.

For these reasons the cell32 is split in two, one for signed and one for
unsigned computations, and the default cell size is now 24 bits.